astock_requirements.txt
# AlphaGPT A股版本依赖
# Python 3.10+ required (recommended: 3.11)
#
# Installation:
#   pip install -r astock_requirements.txt

# ===== 核心依赖 =====

# 深度学习
torch>=2.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.10

# 异步
aiohttp>=3.9.0

# 环境配置
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0
tqdm>=4.66.0

# ===== XtQuant (QMT) =====
# 需要从迅投官方获取安装包
# pip install xtquant  # 或从 QMT 安装目录复制

# ===== 可视化 (可选) =====
streamlit>=1.28.0
plotly>=5.17.0

# ===== 其他 =====
schedule>=1.2.0  # 定时任务

